Hook up scripts to assemble sub-tasks
Previously, we hooked up our custom scripts as dependencies for the
"assemble" task, which meant they got executed only when the main
assemble task was executed, i.e. both the Debug and Release .apk were
assembled. This commit makes it so the scripts are hooked up to the
"assembleDebug" and "assembleRelease" tasks, instead.